Moved daata related endpoints into different 'Blueprints' and referenced from main + renamed gmail
